backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg[binary]
pyotp
python-multipart
email-validator
